Create Scavenge Frame Index in determineCalleeSaves for Scavenge register spilling
Scavenge Register is the temp register we may need after register
allocation.

E.g.
    store {FI#20}

    FI#20 will extend to (SP + Offset) after register allocation

    If Offset exceed 12-bit signed which can't generate by single RISCV
    instruction, we may need a temp register to expand the instruction.

To make sure that we can find a Scavenge register, we have to
reserve the space for spilling.

If the estimated stack offset may exceed 12-bit, stack base relative
access may need scavenge register.
In this case we'll reserve the space for scavenge register spilling.