[loop cannonicalize:] loop cannonicalization for xloops added
In this commit I have added RISCVXloops pass for the backend
which checks the unordered_for metadata attached by clang,
and cannonicalizes the loop for jump to foru: and replaces
bne instruction with a for_u instruction. Right now Loop
Strength Reduction is not supported so one has to use
--mllvm --disable-lsr while compiling. This commit also needs
that the clang has the support for handling the pragma
unordered_for. The changes made are:

-- Added FOR_U and FOR_U64 in the llvm RISC-V backend
-- Added RISCVXloops pass

To test the changes type:

% make compile
% riscv-unknown-elf-objdump -dC matrix-mul-xloops \
    > matrix-mul-xloops.dump

You should be able to see the cannonicalized loop and for_u inst.